inputs to tod_sim_noise_timestreamp must be float64

+++ src/toast/tod/tod_math.py
@@ -221,8 +221,8 @@ def sim_noise_timestream(
             rate,
             firstsamp,
             oversample,
-            freq,
-            psd,
+            freq.astype(np.float64),
+            psd.astype(np.float64),
             tdata,
         )
         return tdata.array()
